Apply suggestions from code review

+++ source/solvers/heat_transfer.cc
@@ -690,9 +690,9 @@ HeatTransfer<dim>::attach_solution_to_output(DataOut<dim> &data_out)
     this->simulation_parameters.physical_properties_manager
       .get_thermal_conductivity_vector();
 
-  const double n_fluids = this->simulation_parameters
+  const unsigned int n_fluids = this->simulation_parameters
                             .physical_properties_manager.get_number_of_fluids();
-  const double n_solids = this->simulation_parameters
+  const unsigned int n_solids = this->simulation_parameters
                             .physical_properties_manager.get_number_of_solids();
 
   // Postprocess heat fluxes
